requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
Django==3.2.3
django-embed-video==1.4.0
filelock==3.0.12
idna==2.10
pipenv==2020.11.15
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
virtualenv==20.4.4
virtualenv-clone==0.5.4
